requirements/dev.txt
autopep8==1.3.3
Django==1.11.8
django-allauth==0.34.0
django-jet==1.0.6
mysqlclient==1.3.12
pep8==1.7.1
PyHamcrest==1.9.0
radish-bdd==0.8.0
selenium==3.8.0
